remove some warnings
git-svn-id: http://192.168.20.11/sequoiadb/trunk@14249 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/mig/migImport.cpp
@@ -483,7 +483,7 @@ INT32 migImport::_run ( INT32 &total, INT32 &succeed )
          // empty bson
          continue ;
       }
-      if ( ( sumSize + bsonSize ) >=
+      if ( (UINT32)( sumSize + bsonSize ) >=
             ( SDB_MAX_MSG_LENGTH - sizeof( MsgOpInsert ) ) )
       {
          sumSize = bsonSize ;
